requirements.txt
streamlit>=1.23.0
pandas
Pillow
altair<5